% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{avana}
\alias{avana}
\title{Pre-computed gRNA efficacy estimates for the Avana library.
Two vectors are provided, both sorted according to the gRNA sequence.
\itemize{
  \item x: Estimated gRNA efficacy for each gRNA. x = 1 means the guide works roughly as an average gRNA would. x = 0 means guide is ineffective.
  \item sdx: Standard deviation of x estimate.
}}
\format{Two numeric vectors.}
\source{
\url{http://www.diamondse.info/}
}
\description{
Pre-computed gRNA efficacy estimates for the Avana library.
Two vectors are provided, both sorted according to the gRNA sequence.
\itemize{
  \item x: Estimated gRNA efficacy for each gRNA. x = 1 means the guide works roughly as an average gRNA would. x = 0 means guide is ineffective.
  \item sdx: Standard deviation of x estimate.
}
}
\author{
Leopold Parts \email{lp2@sanger.ac.uk}
}
\keyword{data}
