% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calc_logfc}
\alias{calc_logfc}
\title{Calculate replicate-averaged log2-fold changes in gRNA frequency.}
\usage{
calc_logfc(expt, reference = "CTRL", condition = "D21")
}
\arguments{
\item{expt}{\code{\link{SummarizedExperiment}} of raw read counts, or replicate-averaged log2-scale gRNA counts}

\item{reference}{String corresponding to factor level in colData$Condition that specifies the change baseline. Default "CTRL".}

\item{condition}{String corresponding to factor level in colData$Condition that specifies the condition for which change is calculated. Default "D21".}
}
\value{
\code{List} object endowed with inferred posterior distributions for one gene.
}
\description{
\code{calc_logfc()} takes a \link{SummarizedExperiment}, which must pass either the \link{validate_logf_table}
or \link{validate_count_table} check, and calculates the replicate-averaged log2-fold change of gRNA frequency
in treatment condition compared to control. The experiment object is endowed with fields \code{logfc_mean} and
\code{logfc_sd} accordingly. If counts are provided, the frequency only averages \code{logf_mean} and \code{logf_sd}
are calculated and stored first as well.
}
