% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{infer_jacks_gene}
\alias{infer_jacks_gene}
\title{Infer JACKS decomposition for a single gene}
\usage{
infer_jacks_gene(data, data_err, n_iter = 5, tol = 0.1, mu0_x = 1,
  var0_x = 1, mu0_ws = 0, var0_w = 10000, tau_prior_strength = 0.5,
  ref_x1 = NA, ref_x2 = NA)
}
\arguments{
\item{data}{\code{\link{matrix}} of replicate-averaged gRNA count log2-fold changes in treatment compared to control for all the gRNAs and experiments for one gene}

\item{data_err}{\code{\link{matrix}} standard deviation of the values in \code{data}.}

\item{n_iter}{An integer number of iterations of JACKS inference performed. Default 5.}

\item{tol}{Tolerance of model fit. Inference stops when lower bound from consecutive rounds does not improve by at least this much. Default 0.1}

\item{mu0_x}{Prior mean of gRNA efficacy \emph{x}. 1 means we assume gRNAs to work well on average. Default 1.}

\item{var0_x}{Prior variance of gRNA efficacy \emph{x}. 1 means we assume gRNAs to have a broad range of efficacies. Default 1.}

\item{mu0_ws}{Prior mean of gene essentialities \emph{w}. 0 means we assume genes to have no effect on average when knocked out. Default 0.}

\item{var0_w}{Prior variance of gene essentiality \emph{w}. Large value means we do not constrain the inferred values. Default 10000.}

\item{tau_prior_strength}{Strength of prior on \emph{tau}. 0.5 means we weight information about uncertainty equally from measurement noise and accuracy of the JACKS decomposition. Default 0.5.}

\item{ref_x1}{Reference gRNA efficacy E(\emph{x}). If provided, will be used directly, instead of inferring from data. Default NA.}

\item{ref_x2}{Reference gRNA efficacy E(\emph{x}^2). If provided, will be used directly, instead of inferring from data. Default NA.}
}
\value{
\code{List} object endowed with inferred posterior distributions for one gene.
}
\description{
\code{infer_jacks_gene()} takes the mean and standard deviation of gRNA log-fold change for one gene across
many experiments, as well as inference parameters (e.g. prior distribution specification),
and performs inference.
A list of posterior distributions is returned, one each for \emph{w}, \emph{x}, \emph{tau}. the estimated gene essentialities, their uncertainties, and directional p-values
are stored in experiment metadata, the gRNA efficacies with uncertainties in
rowData, and prediction of the decomposition and its uncertainty as assays.
}
