% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{jacks_w_gene}
\alias{jacks_w_gene}
\title{Extract JACKS output for a gene}
\usage{
jacks_w_gene(expt, gene)
}
\arguments{
\item{expt}{Output of JACKS inference - SummarizedExperiment endowed with posteriors.}

\item{gene}{Gene name to extract information for.}
}
\value{
\code{data.table} object estimated means and standard deviations of gene essentiality.
}
\description{
\code{jacks_w_gene()} takes JACKS output and gene name, and returns a \code{data.frame}
that has columns for statistics of JACKS inferred posterior for gene essentiality w - mean
and standard deviation. Each row is one cell line, name in row.names.
}
