% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_jacks}
\alias{plot_jacks}
\title{Plot JACKS result for one gene.}
\usage{
plot_jacks(input, gene, lines = NULL, all_logfc = NULL, do_save = TRUE)
}
\arguments{
\item{input}{A \link{SummarizedExperiment} object returned by \code{\link{infer_jacks}} or a list of posteriors returned by \code{\link{infer_jacks_gene}}.}

\item{gene}{String identifying the gene for which to plot results}

\item{lines}{Vector of cell line (experiment) names used. Only used when list of posteriors without annotation is given as input. Default NULL.}

\item{all_logfc}{Vector of all log-fold changes in the experiment. Only used when list of posteriors without annotation is given as input. Default NULL.}

\item{do_save}{Logical scalar of whether to save a .pdf of the output. Default T.}
}
\value{
nothing. Execution stops if errors are found.
}
\description{
\code{plot_jacks()} plots the JACKS decomposition of gRNA log2-fold changes for one gene across
several experiments into an experiment-specific gene effect, and experiment-independent gRNA efficacy.
Input can either be a \link{SummarizedExperiment} object returned by \code{\link{infer_jacks}} or
a list of posteriors returned by \code{\link{infer_jacks_gene}}.
}
