% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_count_table_from_spec}
\alias{read_count_table_from_spec}
\title{Read gRNA counts given sample and gene specification}
\usage{
read_count_table_from_spec(sample_spec, gene_spec, count_prior = 32,
  normalization = "median", window = 800)
}
\arguments{
\item{sample_spec}{Data frame of at least three columns: "Filename": input count file, "Replicate": column names in count file, "Sample": corresponding sample ID to allow combining of replicates within samples.}

\item{gene_spec}{Data frame of at least two columns: gRNA ID (1st col) and corresponding gene (2nd col)}

\item{count_prior}{Scalar double pseudocounts for each gRNA added to observations. Default: 32.}

\item{normalization}{String per-sample normalization method. Default: 'median' [sets median log2-scale count to 0]}

\item{window}{Scalar integer smoothing window used in standard deviation estimation. Default: 800.}
}
\value{
\code{\link{SummarizedExperiment}} object of log-scale sample read counts and estimated standard deviations across replicates
}
\description{
\code{read_count_table_from_spec()} takes a sample specification, and returns a \code{\link{SummarizedExperiment}}
object of read counts.
}
\details{
The sample specification table provides names of count files to read, columns in those files to retain, and the samples these correspond to.
For example,
\code{}
JACKS assumes that each count file has the same number of rows ordered in the same way, corresponding to the same gRNAs.
Every count file has one header line, and the rest of the lines give raw gRNA read counts for one gRNA each, with each column being a different measurement.
Only columns with headers as given in the "Replicate" (2nd column of the specification) are stored from each count file.
Count file columns that map to the same "Sample" (3rd column of the specification) are treated as replicate measurements for that sample.
}
