% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_counts_from_spec_files}
\alias{read_counts_from_spec_files}
\title{Read gRNA counts given sample and gene specification files}
\usage{
read_counts_from_spec_files(count_file, sample_spec_file, replicate_col,
  sample_col, reference_sample, gene_spec_file, grna_col, gene_col,
  count_prior = 32, normalization = "median", window = 800)
}
\arguments{
\item{count_file}{String path to read count file. See \code{read_count_table_from_spec} for requirements.}

\item{sample_spec_file}{String path to sample specification file. Has to contain \code{replicate_col} and \code{sample_col} columns.}

\item{replicate_col}{String Name of the column in \code{sample_spec_file} that contains identifiers for each experimental replicate, as used as the column header for that replicate in the corresponding count file (globally unique identifiers per line, treatment, replicate, etc., not just '1','2', 'A', or 'R1').}

\item{sample_col}{String Name of the column in \code{sample_spec_file} that contains the sample identifiers. For example, if three columns in the count file are replicate measurements of the same sample, they should all have the same sample id here.}

\item{reference_sample}{String Sample name of the single reference sample all other samples will be compared against}

\item{gene_spec_file}{String path to gRNA-gene link specification file. Has to contain \code{grna_col} and \code{gene_col} columns.}

\item{grna_col}{String Name of the column in \code{gene_spec_file} that contains the gRNA identifiers.}

\item{gene_col}{String Name of the column in \code{gene_spec_file} that contains the gene identifiers (to identify which gRNAs map to the same gene).}

\item{count_prior}{Scalar double pseudocounts for each gRNA added to observations. Default: 32.}

\item{normalization}{String per-sample normalization method. Default: 'median' [sets median log2-scale count to 0]}

\item{window}{Scalar integer Length of smoothing window used in standard deviation estimation. Default: 800.}
}
\value{
\code{\link{SummarizedExperiment}} object of log2-scale read count changes compared to reference, and estimated standard deviations across replicates
}
\description{
\code{read_counts_from_spec_files()} takes a file containing raw read counts, a replicate to sample specification file, and a gRNA ID to gene specification file, and returns a \code{\link{SummarizedExperiment}}
object of normalized log2-scale read counts and standard deviations.
}
