% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{run_tests}
\alias{run_tests}
\title{Run tests}
\usage{
run_tests()
}
\value{
nothing. Execution stops if errors are found.
}
\description{
\code{run_tests()} runs all defined unit tests. So far, these include inference from simulations,
real data, and using a reference gRNA efficacy definition; as well as concordance of
python and R versions of JACKS.
}
