% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{validate_count_table}
\alias{validate_count_table}
\title{Validate count table}
\usage{
validate_count_table(counts)
}
\arguments{
\item{counts}{\code{\link{SummarizedExperiment}} object}
}
\value{
TRUE if JACKS can be run on this object; exception otherwise
}
\description{
\code{validate_count_table()} takes in an object, and confirms whether it meets the
requirements to be used in JACKS. A valid count table is of type \code{\link{SummarizedExperiment}},
which includes an assay named \code{'counts'} that contains the raw read counts.
To enable gene-level analysis, the \code{rowData} has to contain a column named \code{'gene'}
that will be used to collage gRNA information. \code{ColData} of the experiment should
be a valid sample table
}
