% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{validate_logf_table}
\alias{validate_logf_table}
\title{Validate log-scale count table}
\usage{
validate_logf_table(logf, require_0median = FALSE)
}
\arguments{
\item{logf}{\code{\link{SummarizedExperiment}} object}

\item{require_0median}{Whether to require per-sample frequencies to have median of 0. Default TRUE.}
}
\value{
TRUE if JACKS can be run on this object; exception otherwise
}
\description{
\code{validate_logf_table()} takes in an object, and confirms whether it meets the
requirements to be used in JACKS. A valid log scale count table is of type SummarizedExperiment,
which includes an assay named 'logf_mean' that contains the log-scale read counts, normalised
and averaged across replicates, as well as an assay 'logf_sd', which contains the standard deviation
of the log-scale counts across replicates. To enable gene-level analysis, the rowData has to contain
a column named 'gene' that will be used to collage gRNA information. ColData of the experiment should
be a valid sample table. Finally, if desired, it is checked that log-fold changes have zero median.
}
