% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{validate_logfchange_table}
\alias{validate_logfchange_table}
\title{Validate log-fold change table}
\usage{
validate_logfchange_table(logfc)
}
\arguments{
\item{logfc}{\code{\link{SummarizedExperiment}} object}
}
\value{
TRUE if JACKS can be run on this object; exception otherwise
}
\description{
\code{validate_logfchange_table()} takes in an object, and confirms whether it meets the
requirements to be used in JACKS. A valid log-fold change table is of type SummarizedExperiment,
which includes an assay named 'logfc_mean' that contains the log-fold change, combined across replicates,
as well as an assay 'logfc_sd', which contains the standard deviation of this estimate.
To enable gene-level analysis, the rowData has to contain a column named 'gene'
that will be used to collage gRNA information. ColData of the experiment should be a valid sample table.
}
