% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{validate_sample_table}
\alias{validate_sample_table}
\title{Validate sample table}
\usage{
validate_sample_table(samples, require_name = TRUE, require_cond = TRUE,
  require_rep = TRUE)
}
\arguments{
\item{samples}{\link{data.frame} of samples}

\item{require_name}{A logical scalar whether 'Name' field is required. Default: TRUE}

\item{require_cond}{A logical scalar whether 'Condition' field is required. Default: TRUE}

\item{require_rep}{A logical scalar whether 'Replicate' field is required. Default: TRUE}
}
\value{
TRUE if JACKS can be run using this object as sample metadata; exception otherwise
}
\description{
\code{validate_sample_table()} takes in an object, and confirms whether it meets the
requirements to be used in JACKS. A valid sample table has one row per sample, and columns
according to the annotation. Depending on strictness of definition, a valid table must have
metadata columns for \code{'Name'} (with capital N; sample name), \code{'Condition'} (condition the sample was
measured in, e.g. \code{"CTRL"}, or \code{"Day21-Puromycin"}), and \code{'Replicate'} (replicate index, e.g. \code{'1'},
\code{'Rep1'}, or \code{'A'}). Samples with same name and condition will be collated across all unique replicates;
different conditions can later be specified to be contrasted in JACKS analysis. If samples are adequately
described, they can be automatically analysed from raw counts; if not, the user has to manually
calculate log-fold changes before running JACKS.
}
